refactor: move PartialAssignmentUtils inside model and rename
